# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTM AI Analytics Dashboard API",
    "description": "제조업 B2B 마케팅 분석을 위한 GA4 기반 대시보드 API",
    "version": "1.0.0",
    "contact": {
      "name": "RTM AI Team",
      "email": "namgyuyooo@rtm.ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://ga-git-main-namgyuyooos-projects.vercel.app",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/dashboard/overview": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "대시보드 개요 데이터 조회",
        "description": "GA4 데이터를 기반으로 한 대시보드 개요 정보를 반환합니다.",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "분석 기간 (예: 7daysAgo, 30daysAgo)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "7daysAgo",
              "enum": ["7daysAgo", "30daysAgo", "90daysAgo"]
            }
          },
          {
            "name": "propertyId",
            "in": "query",
            "description": "GA4 속성 ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "464147982"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "시작 날짜 (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "종료 날짜 (YYYY-MM-DD)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOverview"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analytics/keyword-cohort": {
      "get": {
        "tags": ["Analytics"],
        "summary": "키워드 코호트 분석 데이터 조회",
        "description": "검색 키워드별 코호트 분석 데이터를 반환합니다.",
        "parameters": [
          {
            "name": "propertyId",
            "in": "query",
            "description": "GA4 속성 ID",
            "required": false,
            "schema": {
              "type": "string",
              "default": "464147982"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "분석 기간",
            "required": false,
            "schema": {
              "type": "string",
              "default": "7daysAgo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordCohortResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DashboardOverview": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "kpis": {
                "type": "object",
                "properties": {
                  "totalSessions": {
                    "type": "integer",
                    "description": "총 세션 수"
                  },
                  "totalUsers": {
                    "type": "integer",
                    "description": "총 사용자 수"
                  },
                  "pageViews": {
                    "type": "integer",
                    "description": "총 페이지뷰"
                  },
                  "conversions": {
                    "type": "integer",
                    "description": "총 전환 수"
                  },
                  "conversionRate": {
                    "type": "number",
                    "format": "float",
                    "description": "전환율"
                  }
                }
              },
              "topCampaigns": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TopCampaign"
                }
              },
              "topPages": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TopPage"
                }
              }
            }
          }
        }
      },
      "TopCampaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "캠페인명"
          },
          "source": {
            "type": "string",
            "description": "트래픽 소스"
          },
          "medium": {
            "type": "string",
            "description": "매체"
          },
          "sessions": {
            "type": "integer",
            "description": "세션 수"
          },
          "conversions": {
            "type": "integer",
            "description": "전환 수"
          }
        }
      },
      "TopPage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "페이지 경로"
          },
          "title": {
            "type": "string",
            "description": "페이지 제목"
          },
          "pageViews": {
            "type": "integer",
            "description": "페이지뷰 수"
          },
          "users": {
            "type": "integer",
            "description": "사용자 수"
          },
          "avgSessionDuration": {
            "type": "number",
            "format": "float",
            "description": "평균 세션 시간"
          },
          "bounceRate": {
            "type": "number",
            "format": "float",
            "description": "바운스율"
          }
        }
      },
...
      "KeywordCohortResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "keyword": {
                  "type": "string",
                  "description": "검색 키워드"
                },
                "impressions": {
                  "type": "integer",
                  "description": "노출 수"
                },
                "clicks": {
                  "type": "integer",
                  "description": "클릭 수"
                },
                "ctr": {
                  "type": "number",
                  "format": "float",
                  "description": "클릭률"
                },
                "position": {
                  "type": "number",
                  "format": "float",
                  "description": "평균 검색 순위"
                },
                "conversions": {
                  "type": "integer",
                  "description": "전환 수"
                },
                "revenue": {
                  "type": "number",
                  "format": "float",
                  "description": "매출"
                }
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "오류 메시지"
          },
          "details": {
            "type": "string",
            "description": "상세 오류 정보"
          }
        }
      }
    }
  }
}